fix compiler warnings, fix implicit conversion of numeric types, refactor out PhysicsBody
use size_t to match container size()
fix floats initialized with double literals

refactor/move PhysicsBody mixin into GObject
initializeRadar: declare base definition inline

pragma ignore: Visual C++ inherited via dominance warning
chipmunk.h fix _USE_MATH_DEFINES redefintion
App instance instantiated in main: rename to inst to avoid name collision with global app pointer.

.gitignore: VS code analysis files
VS libcocos2d: linker option /ignore:4099 - ignore lack of pdb (debug symbols databse files) for prebuilt libraries.

+++ Classes/AI.hpp
@@ -215,7 +215,7 @@ public:
 	virtual void update(StateMachine& fsm);
 protected:
 	Path path;
-	int currentTarget = 0;
+	size_t currentTarget = 0;
 	bool loop = false;
 };
